thermosphere: attempt to output to uart-c

+++ thermosphere/src/utils.h
@@ -26,7 +26,7 @@
 #define MASK2(a,b)  (MASK(a) & ~MASK(b))
 #define MASK2L(a,b) (MASKL(a) & ~MASKL(b))
 
-#define MAKE_REG32(a)   (*(volatile u32 *)(a))
+#define MAKE_REG32(a)   (*(volatile u32 *)(uintptr_t)(a))
 
 #define ALIGN(m)        __attribute__((aligned(m)))
 #define PACKED          __attribute__((packed))
